refactor: Move the `sync_restriction` to `restrictions/sync`
