+++ tools/serverpod_cli/lib/src/analyzer/models/validation/restrictions/sync.dart
