+++ tools/serverpod_cli/lib/src/database/create_definition.dart
@@ -1,6 +1,6 @@
 import 'package:serverpod_cli/src/analyzer/models/definitions.dart';
 import 'package:serverpod_cli/src/analyzer/models/utils/quote_utils.dart';
-import 'package:serverpod_cli/src/analyzer/models/validation/sync_restrictions.dart';
+import 'package:serverpod_cli/src/analyzer/models/validation/restrictions/sync.dart';
 import 'package:serverpod_cli/src/config/config.dart';
 import 'package:serverpod_cli/src/generator/types.dart';
 import 'package:serverpod_database/serverpod_database.dart';
